Update section on embed ipywidgets
Updated code example by initialising the embedded ipywidget of the custom class within the __init__ function, so that each instance of the custom class has its own instance of the ipywidget. Also added the possibility of customising the embedded widget upon instantiating the custom class, by using keyword arguments of the _init__function.
Harmonised quote style in the code example.