refactor: only apply nest_asyncio.patch() when needed
We always apply nest_asyncio.patch(), which does not work when a
different event loop is installed.
Instead, we only need to apply when we use the FileInput widget.
Also, we *can* only apply this route when we are using the ipython
kernel, since only that kernel support this API:

        while not ipython.kernel.msg_queue.empty():
            await ipython.kernel.do_one_iteration()