fix: no hard dependency on nest_asyncio, use only when needed.
Only when we have a current event loop do we need nest_asyncio to be
used and installed. In case where readinto is called from a thread
(this is something we do in solara) which does not have an event loop
yet, we should not need nest_asyncio.

When needed, and not installed, we raise an error with an informative
installation instruction.