feat: improve Vuetify CSS scoping
Previously the CSS was simply scoped to `.vuetify-styles` in `@mariobuikhuizen/vuetify` JS-package. We moved the scoping to be here to be closer to the rest of the relevant code. The scoping is now also done based using the `:where` CSS selector, since this has no impact on specificity, meaning that Vuetify CSS should be less likely to override other styles unexpectedly.